+++ fixtures/remix-1.18.0/express/app/routes/$teamId.$projectId.overview.tsx
@@ -1,0 +1,3 @@
+export default function BirdRoute() {
+  return <div>$teamId.$projectId.overview</div>;
+}