+++ fixtures/remix-1.18.0/express/app/routes/throw-redirect.tsx
@@ -1,0 +1,9 @@
+import { redirect } from "@remix-run/server-runtime";
+
+export function loader() {
+  throw redirect("/");
+}
+
+export default function ClientErrorRoute() {
+  return <div>throw-redirect</div>;
+}