+++ fixtures/remix-1.18.0/express/app/routes/throw-response.tsx
@@ -1,0 +1,7 @@
+export function loader() {
+  throw new Response("throw-response", { status: 400 });
+}
+
+export default function ClientErrorRoute() {
+  return <div>throw-response</div>;
+}