+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @metronome-sh/cli
 
+## 7.1.0-next.10
+
+### Patch Changes
+
+- Fixed routePath
+
 ## 7.1.0-next.9
 
 ### Patch Changes